ai-service/requirements.txt
fastapi>=0.115.2,<1.0
uvicorn[standard]>=0.34.0,<1.0
pymupdf>=1.26.6,<2.0
openai>=1.52.2,<2.0
python-multipart>=0.0.18,<1.0
pydantic>=2.11.0,<3.0
python-dotenv>=1.0.1,<2.0
